fix(css): Fix Styling for CM Tooltip Glossary
The CM Tooltip Glossary overrides the text color for acronyms to be
always "dark". This causes issues, when the text is meant to be white,
for example, on dark background (like on the banner image on our
homepage.

Now we reset the text color to be taken from the context.

Also, the two-column layout for the glossary overview is now only
applied on desktop monitors. This should calm issues that were reported,
when the glossary is opened on mobile phones, where sometimes the
tooltips "went wild" and were displayed at random locations.