fix(css): Selected Day Coloring on Mobile Devices
On mobile devices, if you select a day, a dark blue box is display for
the selected day. The text still default to black color which makes
overall a black text appear on very dark blue background.

This change fixes it, by being applied in mobile scenarios where we now
switch to white text color instead.