fix(css): Fixes WPAdminBar Event Actions
Event actions were hidden, because they appear in the "overflow" area of
the administrative bar. This collided with this rule:

```css
#wpadminbar { overflow: hidden; }
```

So, just to make something "pretty" we broke the events calendar layout.

Just removing this rule, to repair it. Note, that the original approach
had a comment "Hide overflow, as wrapping looks bad", so that we may
want to check that. On the other hand, we do not expect administrative
actions to be done a lot on mobile devices.